Chore: Move Legacy models into correct package
